[FIX] 修改删除路由接口

+++ apix-service/service-api-management/src/main/java/com/sinosdx/service/management/service/impl/AppApiGatewayServiceImpl.java
@@ -210,6 +210,9 @@ public class AppApiGatewayServiceImpl implements AppApiGatewayService {
         Set<String> updateGatewayIdSet = new HashSet<>();
         for (Api api : apiList) {
             List<ApplicationApiGateway> gatewayList = appApiGatewayMapper.queryListByCondition(null, applicationId, api.getId(), appClientCode);
+            if (CollectionUtils.isEmpty(gatewayList)) {
+                continue;
+            }
             String gatewayId = gatewayList.get(0).getGatewayId();
             for (ApplicationApiGateway apiGateway : gatewayList) {
                 appApiGatewayMapper.deleteById(apiGateway);
